Add Algorithm interface.
